Moving request_proxy_list to shared
Needs used in both scrapers.py and integration.py
